Linux build fix.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@6686 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/VideoCommon/Src/BPFunctions.cpp
@@ -187,7 +187,7 @@ void OnPixelFormatChange(const BPCmd &bp)
 		default:
 			break;
 	}
-	if (convtype == (unsigned int)-1)
+	if (convtype == -1)
 	{
 		PanicAlert("Unhandled EFB format change: %d to %d\n", old_format, new_format);
 		goto skip;
